v1.1.3

+++ examples/CheezPPG_OLED/CheezPPG_OLED.ino
@@ -86,8 +86,8 @@ void loop() {
       display.clearDisplay();
     }
 
-    HR = czPPG.getPpgHr();
-    HRV = czPPG.getPpgHrv();
+    HR = (int)czPPG.getPpgHr();
+    HRV = (int)czPPG.getPpgHrv();
     bpPPG = czPPG.getFilterPPG();
 
   }
